+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to
   - [#2315](https://github.com/iovisor/bpftrace/pull/2315)
 - Add %rh option to print buffer as hex without \x
   - [#2445](https://github.com/iovisor/bpftrace/pull/2445)
+- Add stdbool.h to built-in headers
+  - [#2380](https://github.com/iovisor/bpftrace/pull/2380)
 #### Changed
 - Raise minimum versions for libbpf and bcc and vendor them for local builds
   - [#2369](https://github.com/iovisor/bpftrace/pull/2369)